+++ tests/test_loggers.py
@@ -14,7 +14,14 @@
 
 from __future__ import absolute_import, division, print_function
 
-from structlog._loggers import PrintLogger, ReturnLogger
+import sys
+
+from structlog._loggers import (
+    PrintLogger,
+    PrintLoggerFactory,
+    ReturnLogger,
+    ReturnLoggerFactory,
+)
 
 
 def test_return_logger():
@@ -42,3 +49,47 @@ class TestPrintLogger(object):
         assert repr(PrintLogger()).startswith(
             "<PrintLogger(file="
         )
+
+
+class TestPrintLoggerFactory(object):
+    def test_does_not_cache(self):
+        """
+        Due to doctest weirdness, we must not re-use PrintLoggers.
+        """
+        f = PrintLoggerFactory()
+        assert f() is not f()
+
+    def test_passes_file(self):
+        """
+        If a file is passed to the factory, it get passed on to the logger.
+        """
+        l = PrintLoggerFactory(sys.stderr)()
+        assert sys.stderr is l._file
+
+    def test_ignores_args(self):
+        """
+        PrintLogger doesn't take positional arguments.  If any are passed to
+        the factory, they are not passed to the logger.
+        """
+        PrintLoggerFactory()(1, 2, 3)
+
+
+class TestReturnLoggerFactory(object):
+    def test_builds_returnloggers(self):
+        f = ReturnLoggerFactory()
+        assert isinstance(f(), ReturnLogger)
+
+    def test_caches(self):
+        """
+        There's no need to have several loggers so we return the same one on
+        each call.
+        """
+        f = ReturnLoggerFactory()
+        assert f() is f()
+
+    def test_ignores_args(self):
+        """
+        ReturnLogger doesn't take positional arguments.  If any are passed to
+        the factory, they are not passed to the logger.
+        """
+        ReturnLoggerFactory()(1, 2, 3)